requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
Automat==20.2.0
backcall==0.2.0
baidu-aip==2.2.18.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
constantly==15.1.0
cryptography==3.2.1
cssselect==1.1.0
datacompy==0.7.1
decorator==4.4.2
et-xmlfile==1.0.1
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.0.0
incremental==17.5.0
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
itemadapter==0.1.1
itemloaders==1.0.3
jdcal==1.4.1
jedi==0.17.2
jieba==0.42.1
jmespath==0.10.0
joblib==0.17.0
jsonpath==0.82
jupyter-client==6.1.7
jupyter-core==4.6.3
lxml==4.6.1
MouseInfo==0.1.3
numpy==1.19.2
opencv-python==4.4.0.46
openpyxl==3.0.5
packaging==20.4
pandas==1.1.3
parsel==1.6.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
pluggy==1.0.0.dev0
prompt-toolkit==3.0.8
Protego==0.1.16
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.20
PyDispatcher==2.0.5
PyGetWindow==0.0.9
Pygments==2.7.2
PyHamcrest==2.0.2
PyMsgBox==1.0.9
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.6
pytest==6.1.2
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.1
pywin32==228
pyzmq==19.0.2
queuelib==1.5.0
requests==2.24.0
retrying==1.3.3
rpa==1.27.1
scikit-learn==0.23.2
scipy==1.5.4
Scrapy==2.4.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
sklearn==0.0
tagui==1.27.1
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.51.0
traitlets==5.0.5
Twisted==20.3.0
urllib3==1.25.11
w3lib==1.22.0
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
zope.interface==5.1.2
